requirements.txt
aiofiles==0.5.0
alembic==1.6.5
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
autopep8==1.5.7
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.8
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
httptools==0.3.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.5
MarkupSafe==2.0.1
orjson==3.6.1
passlib==1.7.4
praw
promise==2.3
psycopg2==2.9.1
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
six
SQLAlchemy==1.4.23
starlette==0.14.2
textblob
toml==0.10.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.13.4
watchgod==0.7
websockets==8.1